pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ercot_scraping"
version = "0.1.0"
dependencies = [
    "pandas",
    "requests",
    "python-dotenv",
    "ratelimit",
    "chardet",
    "aiosqlite",
    "aiohttp",
    "tqdm",
    "pydantic"
]
requires-python = "<=3.11.9 "

[project.scripts]
ercot = "ercot_scraping.run:main"  # Changed from fetch_prices to ercot

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cover",  # Updated version
    "coverage",    # Updated version
    "pytest-mock", # Added for better mocking support
    "autopep8",  # Added for code formatting
    "flake8",   # Added for linting
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=ercot_scraping --cov-report=xml --cov-report=term-missing -v"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
branch = true
source = ["ercot_scraping"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/__pycache__/*",
    "*/setup.py",
    "*/conftest.py", 
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass"
]
show_missing = true
skip_covered = true
sort = "Cover"

[tool.flake8]
# Allow lines up to 88 characters (to match Black's default)
max-line-length = 140
